+++ src/endpoints/grpc/proto/mqbridge/bridge.proto
